backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.3.3
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.20
celery==5.4.0
redis==5.2.1
httpx==0.28.1
python-dotenv==1.0.1
anthropic>=0.84.0
langgraph>=1.0.9
langchain-core>=0.3.0
python-docx>=1.1.0
python-pptx>=1.0.0
minio>=7.2.0
